[bugfix] mooseText.c: ensure cursor moves back after backspace
Currently, when we do a backspace in mooseText the cursor does
not move along with backspace key meaning to delete multiple
characters (i.e. a word) you need to use the right arrow key
along with the backspace key otherwise you delete one character
and stop because there is nothing else to delete - from the
perspective of the program.

This commit fixes the issue by ensuring that the cursor position
decrements by 1 everytime a character has been deleted.